+++ ys-client-server-1347/homework/Controllers/Feed/FeedItemInfoTableViewCell.swift
@@ -45,6 +45,6 @@ class FeedItemInfoTableViewCell: UITableViewCell {
             }
         }
         
-        feedItemPostDate.text = postDate.getDateStringFromUTC()
+        feedItemPostDate.text = postDate.getRelativeDateStringFromUTC()
     }
 }
