+++ ys-client-server-1347/homework/Controllers/Feed/FeedTableViewController.swift
@@ -145,16 +145,3 @@ class FeedTableViewController: UITableViewController {
         }
     }
 }
-
-extension Double {
-    func getDateStringFromUTC() -> String {
-        let date = Date(timeIntervalSince1970: self)
-        
-        let dateFormatter = DateFormatter()
-        dateFormatter.locale = Locale(identifier: "ru_RU")
-        dateFormatter.dateStyle = .medium
-        dateFormatter.timeStyle = .short
-        
-        return dateFormatter.string(from: date)
-    }
-}
