+++ ys-client-server-1347/homework/Controllers/Groups/GroupTableViewCell.swift
@@ -40,16 +40,3 @@ class GroupTableViewCell: UITableViewCell {
         }
     }
 }
-
-extension Int {
-    var formatted: String {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .decimal
-        formatter.maximumFractionDigits = 2
-        formatter.locale = Locale(identifier: "RU")
-        
-        let number = NSNumber(value: self)
-        let formattedValue = formatter.string(from: number)!
-        return "\(formattedValue)"
-    }
-}
